requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
arrow==1.3.0
    # via cookiecutter
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
backports-zoneinfo==0.2.1
    # via
    #   -r requirements/base.txt
    #   django
binaryornot==0.4.4
    # via cookiecutter
boto3==1.34.83
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
botocore==1.34.83
    # via
    #   -r requirements/base.txt
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via requests
chardet==5.2.0
    # via binaryornot
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   code-annotations
    #   cookiecutter
code-annotations==1.8.0
    # via -r requirements/test.in
cookiecutter==2.6.0
    # via xblock-sdk
coverage[toml]==7.4.4
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-appconf
    #   django-statici18n
    #   edx-i18n-tools
    #   openedx-django-pyfs
    #   xblock-sdk
django-appconf==1.0.6
    # via
    #   -r requirements/base.txt
    #   django-statici18n
django-statici18n==2.4.0
    # via -r requirements/base.txt
edx-i18n-tools==1.5.0
    # via -r requirements/base.txt
exceptiongroup==1.2.0
    # via pytest
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.1
    # via
    #   -r requirements/base.txt
    #   openedx-django-pyfs
    #   xblock-sdk
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via
    #   code-annotations
    #   cookiecutter
jmespath==1.0.1
    # via
    #   -r requirements/base.txt
    #   boto3
    #   botocore
lazy==1.6
    # via
    #   -r requirements/base.txt
    #   xblock
lxml==5.2.1
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
    #   xblock
    #   xblock-sdk
mako==1.3.3
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-utils
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   mako
    #   xblock
mdurl==0.1.2
    # via markdown-it-py
openedx-django-pyfs==3.6.0
    # via
    #   -r requirements/base.txt
    #   xblock
packaging==24.0
    # via pytest
path==16.14.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pbr==6.0.0
    # via stevedore
pluggy==1.4.0
    # via pytest
polib==1.2.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pygments==2.17.2
    # via rich
pypng==0.20220715.0
    # via xblock-sdk
pytest==8.1.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.in
pytest-django==4.8.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   arrow
    #   botocore
    #   xblock
python-slugify==8.0.4
    # via
    #   code-annotations
    #   cookiecutter
pytz==2024.1
    # via
    #   -r requirements/base.txt
    #   xblock
pyyaml==6.0.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   cookiecutter
    #   edx-i18n-tools
    #   xblock
requests==2.31.0
    # via
    #   cookiecutter
    #   xblock-sdk
rich==13.7.1
    # via cookiecutter
s3transfer==0.10.1
    # via
    #   -r requirements/base.txt
    #   boto3
simplejson==3.19.2
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-sdk
    #   xblock-utils
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   fs
    #   fs-s3fs
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.2.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.11.0
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   rich
urllib3==1.26.18
    # via
    #   -r requirements/base.txt
    #   botocore
    #   requests
web-fragments==2.2.0
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-sdk
    #   xblock-utils
webob==1.8.7
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-sdk
xblock[django]==3.1.0
    # via
    #   -r requirements/base.txt
    #   xblock-sdk
    #   xblock-utils
xblock-sdk==0.7.0
    # via -r requirements/test.in
xblock-utils==4.0.0
    # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
